% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcmodifydb-package.R
\docType{package}
\name{dcmodifydb-package}
\alias{dcmodifydb}
\alias{dcmodifydb-package}
\title{dcmodifydb: Modifying rules on db}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Edwin de Jonge \email{edwindjonge@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
