% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_con.R
\name{get_table_con}
\alias{get_table_con}
\title{Get table connection}
\usage{
get_table_con(table, con = NULL, copy = NULL)
}
\arguments{
\item{table}{either a character or a tbl_sql object}

\item{con}{dbi connection}

\item{copy}{should a copy of the table be generated?}
}
\description{
Gets a table connection, possibly a copy of the table.
}
\keyword{internal}
