% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcmodifydb-package.R
\docType{package}
\name{dcmodifydb-package}
\alias{dcmodifydb}
\alias{dcmodifydb-package}
\title{dcmodifydb: Modifying Rules on a DataBase}
\description{
Apply modification rules from R package 'dcmodify' to the database, prescribing and documenting deterministic data cleaning steps on records in a database. The rules are translated into SQL statements using R package 'dbplyr'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/data-cleaning/dcmodifydb}
  \item Report bugs at \url{https://github.com/data-cleaning/dcmodifydb/issues}
}

}
\author{
\strong{Maintainer}: Edwin de Jonge \email{edwindjonge@gmail.com} (\href{https://orcid.org/0000-0002-6580-4718}{ORCID})

Other contributors:
\itemize{
  \item Wytze Gelderloos [contributor]
}

}
\keyword{internal}
