% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{dump_sql}
\alias{dump_sql}
\title{Write generated sql}
\usage{
dump_sql(x, table, con = NULL, file = stdout(), ...)
}
\arguments{
\item{x}{\code{dcmodify::modifier()} object with rules to be written}

\item{table}{either a \code{\link[dplyr:tbl]{dplyr::tbl()}} object or a \code{character} with table name}

\item{con}{optional, when \code{table} is a character a dbi connection.}

\item{file}{to which the sql will be written.}

\item{...}{not used}
}
\description{
Writes generated sql to file
}
