% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_working.R
\name{is_working_db}
\alias{is_working_db}
\title{Check if UPDATE statement is functional}
\usage{
is_working_db(updates, tab, n = 2)
}
\arguments{
\item{updates}{\code{list} of update statements object (\code{\link[=modifier_to_sql]{modifier_to_sql()}})}

\item{tab}{tbl object}

\item{n}{number of records to use in this check}
}
\value{
\verb{logical with which statements are working}
}
\description{
Extract UPDATE statements from modifier object.
}
