% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{modifier_to_sql}
\alias{modifier_to_sql}
\title{Extract UPDATE statements}
\usage{
modifier_to_sql(x, table, con = NULL)
}
\arguments{
\item{x}{\code{dcmodify::modifier()} object}

\item{table}{table name}

\item{con}{optional connection}
}
\value{
\code{list} of sql UPDATE statements.
}
\description{
Extract UPDATE statements from modifier object.
}
