% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifier_to_sql.R
\name{modifier_to_sql}
\alias{modifier_to_sql}
\title{Utility function to extract SQL statements}
\usage{
modifier_to_sql(x, table, con = NULL)
}
\arguments{
\item{x}{\code{dcmodify::modifier()} object}

\item{table}{table object}

\item{con}{optional connection}
}
\value{
\code{list} of sql UPDATE statements.
}
\description{
Extract UPDATE statements from modifier object as a list of SQL statements.
A user should normally be using \code{\link[=modify]{modify()}} or \code{\link[=dump_sql]{dump_sql()}}, but this
function may be useful.
}
\seealso{
Other sql translation: 
\code{\link{dump_sql}()}
}
\concept{sql translation}
