% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{modify,ANY,modifier-method}
\alias{modify,ANY,modifier-method}
\title{Test}
\usage{
\S4method{modify}{ANY,modifier}(dat, x, ...)
}
\arguments{
\item{dat}{\code{\link[=tbl_dbi]{tbl_dbi()}} object, table in a DBI database}

\item{x}{\code{\link[dcmodify:modifier]{dcmodify::modifier()}} object.}

\item{...}{unused}
}
\description{
test
}
